pyproject.toml
[project]
name = "relatorio-atual"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.151.0",
    "google-auth-oauthlib>=1.2.1",
    "pandas>=2.2.3",
    "psycopg2>=2.9.10",
    "pyinstaller>=6.11.0",
    "python-dotenv>=1.0.1",
]
